[FP-7564] Write composite fingerprints when all composite data is available in the update (#9629)
## Describe your changes
Here, we're putting aaaaaall of the pieces together from the PRs below this - we _finally_ start writing composite fingerprints! With the caveat that we're only doing it in the simple case where all fields that compose the fingerprint are updated together. In other words, we won't recompute `composite.name_dob` when a data update is made that _only_ updates `id.dob` - i will do that in a follow-up PR.

The general approach involves adding a new kind of `FingerprintScope`: "partial" fingerprints (accepting other name suggestions). Certain DIs need to be finerprinted (aka HMACed in the enclave) solely for the purpose of being later used to compose a composite fingerprint. These "partial" fingerprints might be for lower-cardinality data, so they are never saved to the database

## Caveats

I was waffling between two possible implementations (thread [here](https://onefootprint.slack.com/archives/C04RHCM8FU0/p1715386627101009)): storing either
1. `hmac(k, sha256(first) || sha256(last) || sha256(dob))` - the "HMAC of hashes"
2. `sha256(hmac(k, first) || hmac(k, last) || hmac(k, dob))` - the "hash of HMACs." This option has some length extension vulnerabilities

I ended up going with (2) for the pure purpose of trying to keep the fingerprinting code a little bit more readable. We have two codepaths that fingerprint data: one from plaintext (when vaulting new data) and one from ciphertext (when prefilling a user's data into a new tenant as they one-click onboard). Option (1), the "HMAC of hashes," would require us to get the plaintext in the codepath where we normally generate fingerprints from ciphertext.

I don't feel super strongly about the choice of (2) over (1) - maybe my opinion will change as I implement the next step of generating composite fingerprints when only part of data is updated. But for now, it seems like this option is nicer. It actually doesn't require a single codechange to the codepath for generating fingerprints as we prefill a user's data into a tenant.

<!--- Are there any caveats to this changeset? -->

## Test Plan
Will add more tests in follow-up PR