Differentiate between building fingerprints for new / existing vaults (#9678)
## Describe your changes
All codepaths for updating vault data go through FingerprintedDataRequest. And they can be split into two use cases: adding data to a brand new vault vs adding data to an existing vault. Only in the latter case will we need to do the work of fingerprinting existing ciphertext (for composite fingerprints).

This PR just splits the two callsites - we'll use this in a follow-up PR

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->