[FP-7660] Also make tenant-scoped composite fingerprints for name (#9680)
## Describe your changes
This was much easier to do than i anticipated!

In order to replace some of the hairy code that performs AND filters on multiple fingerprints for the dashboard to search by name, we now can have a composite fingerprint on name!

We wanted this to be tenant-scoped since it's significantly lower cardinality than NameDob

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Locally for now. WIll add coverage to full name search in the dashboard after it's implemented
```
$ curl localhost:8000/users -X POST -d '{"id.first_name": "elliot", "id.last_name": "Forde", "id.dob": "1998-01-04"}' -u sk_test_bO5lzm9r2HLlloOBeizIqmPM3tILO94n1i:
```
```
footprint_db=# select kind, scope from fingerprint where scoped_vault_id ='su_gIEzEEwMezWQD6IKSmeTm7';
        kind        | scope
--------------------+--------
 id.first_name      | tenant
 id.last_name       | tenant
 composite.name_dob | global
 composite.name     | tenant
(4 rows)
```