[billing] Clean up implementation of uncontracted line items (#9755)
## Describe your changes
So silly: we have some code that errors loudly if a tenant uses a
product for which we don't have a price set. We used to need to
configure a "price" for each "product" in stripe that we would use to
make the line item on the invoice when the product is uncontracted. Now,
we just make a freeform line item

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan

I can test this locally! yay

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/c33aef14-e04a-4940-ad1c-6802e8b59c3e.png)