validate we are sending kyb verification checks (#9781)
## Describe your changes
Adds API validation so you must create a verification check for KYB if this is a KYB playbook and you aren't skipping KYB

We cannot set skip_kyb via the API, so i don't have any itests for the opposite case. eventually skip_kyb will be replaced by the absence of the verification check (which will be interesting to validate..)

## Caveats
one rafa merges his change, we can merge this. then we get assurances we're always sending verification checks for KYB and we can start relying on it for app logic

## Test Plan
these are tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->