[DNL] Fix unwraps in middesk request creation (#9778)
## Describe your changes
we have a bunch of [unwraps](https://github.com/onefootprint/monorepo/pull/9686#discussion_r1607346533) currently in the code the translates between data we load from the vault and data we send to middesk.

we need an interface layer to error if we don't have the right data in the vault. this adds that and branches on whether we're configured to do a TinCheck only, or the full KYB using our new VerificationChecks construction

validations are based on Middesk docs as to which fields are required https://docs.middesk.com/reference/create-a-business

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->