[1/4] Start reconciling confusing flags on scoped_vault (#9740)
## Describe your changes
We had a lot of different flags on ScopedVault that control different
things:
- `show_in_search` was used to attempt to hide users that were created
via the signup challenge API and hadn't yet been verified. But we even
forgot to look for this column in a few places and had a bug where we
would show duplicate vaults that were unverified
- `deactivated_at` was used by the old `DELETE /users/<>` API to "soft
delete" a vault - hide it from the dashboard, prevent users from logging
in, etc. I'll keep this column around as timestamp information
- `is_billable` was ~half used in billing queries so that we'd only
charge monthly PII storage on users for which the tenant can decrypt
data. So, we'd essentially make vaults billable as soon as they're
created on the first tenant. But if a user one-clicked onto a new
tenant, it wouldn't be billable until they authorize the workflow. I'm
planning to change this behavior and just make the scoped vault billable
on the second scoped vault - we have such a small volume of one clicks
and i think we have the leeway to decide how we want this to act. The
danger of having a separate column is that we don't actually use it
properly - and in this case, we _weren't_ using it properly for all
billable events

All of these purposes are soooo similar and so easy to misuse. So I'm
planning to drop `show_in_search` and `is_billable` and combine their
purposes into one `is_active` column - a vault that is "active" is
billable, shows up in dashboard search, and can be logged into.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->