Reduce audit event write batch size in bulk decryptions (#9787)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
I don't think audit events and access logs are the primary problem in
the monthly `/vault_proxy/reflect` latency issues, but it might be a
contributing factor.

In large decryption batches (up to 5000 tokens) we currently create
audit & access events in an equally large batch, but there's not a great
reason the whole set of events needs to be created transactionally. If
the request is sent, I think that's reason enough to create an audit
event, even if it timed out due to audit event creation itself.

This PR reduces the batch size for writing audit/access events, so there
will be about 10 shorter transactions for event creation in the largest
batch decryptions. My theory is that keeping transctions open for a
shorter period of time will reduce the cost of these writes.

## Caveats
<!--- Are there any caveats to this changeset? -->
There may be audit/access events created for a "mini-batch" of a batch
decrypt if creating audit/access events for a subsequent "mini-batch"
fails. IMO this is fine.

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->