Support serializing just the obc ID in /validate for composer (#9801)
## Describe your changes
https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1716302995192359

Horrible - composer just decided to use the obc_id to infer alpaca vs
apex, just because it happened to be included in the response...

this was a deprecated field, but we didn't remove it from their view
fast enough. so we have to keep compatibility for their current use
case. maybe we can migrate away from it in the future if we start
exposing playbook keys

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Here are all the tenants on a legacy version of the API - i will bump
composer to version 2 so they continue getting obc_id but none of the
other legacy fields. Then I will merge the PR above this that gets rid
of the backcompat for v1 entirely
```
footprint=> select id, name, pinned_api_version from tenant where pinned_api_version is not null;
             id             |       name       | pinned_api_version
----------------------------+------------------+--------------------
 org_RW4mzhcUJ9mkAg4h2ulXga | Burst            |                  1
 org_AsXQlVLpin64CIjTtW3W1m | mpm792@gmail.com |                  1
 org_qYus55jt4nQCWL2dq5tCcI | Backpack LLC     |                  1
 org_hEjNnwE8KqQE9M6T0ak0JW | culdesac.com     |                  1
 org_e2FHVfOM5Hd3Ce492o5Aat | Footprint Live   |                  1
 org_UjStFCmPcBC9LR3DcXS8ot | Voura            |                  1
 org_hyZP3ksCvsT0AlLqMZsgrI | Acme Inc.        |                  1
 org_9QPsH20xLJCvSRRvxgQtFj | Composer         |                  1
 org_PtnIJT4VR35BS9xy0wITgF | Fractional       |                  1
(9 rows)
```