Support PATCHing billing profile (#9795)
## Describe your changes
Similar to PR below this, but also supports updating billing profile now.

Billing profile has a new twist: fields on the billing profile are nullable, so we need to support (via API) actually clearing out values on the billing profile. To support this, we actually distinguish between (1) explicit null JSON value and (2) unprovided JSON key. The client serializes `undefined` as no key present and `null` as an explicit JSON null. This distinction allows the client to explicitly distinguish between, say, setting price of KYC to `null` vs leaving the price of KYC unchanged (with `undefined`)

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
In PR above this