[Vault DR] Fetch org private key from a script (#9810)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
This PR has a bit of cleanup, but the main objective is changing the
`--org-private-key` flag to `--org-private-key-plugin`. By adding this
layer of indirection for fetching the key, we allow for a wide variety
of storage mechanisms for the private key, and reduce the risk of the
private key being leaked to shell history or left on the filesystem. For
example, the plugin can be a shell script that fetches an AWS SSM
parameter.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->