Allow firm employees to update a tenant's billing profile (#9796)
## Describe your changes
Using the API below this, starts supporting updating a tenant's billing profile, which we used to manually do in the database.

This has a unique twist: the backend actually supports the distinction between `null` values and `undefined` values - `null` means a value should be updated to empty while `undefined `means the value should be left unchanged. On the frontend, we treat a form value of `""` as `null`

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Locally:

https://github.com/onefootprint/monorepo/assets/21066341/93f8103a-6cd0-4730-b2d0-9b9b69615609