[BE-25] Track one-click KYC billing events (#9816)
## Describe your changes
BE-25

Trying to get this in before EOM so next month we can actually bill for these.

## Caveats
This takes the interesting approach of forcing the KYC waterfall to know a little bit about billing in order to bill properly for Kyc or OneClickKyc... which i don't love.

An alternative implementation could make a billing event for OneClickKyc when the Workflow is created (or completed? idk). When making the invoice, the KYC line item can be for the quantity (Kyc - OneClickKyc). But that is also kind of funky. Wdyt?

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->