Fix svix test in remote IT (#9899)
## Describe your changes
These tests used to rely on having access to the same secrets used to
deploy the server, which doesn't make sense. So i think i'll just
disable the webhook test in dev / prod - it will still run for each PR
when it's merged, but just against the local backend server rather than
against the server deployed to dev.

We could have this run against dev too by leaking the svix dev API key
(which i don't mind so much). Or we could have this run in both prod and
dev by having a single hooky API that svix is talking to

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->