Use cargo-chef for caching dependency builds with Docker (#9985)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
Being able to do realistic local builds frees us from most of the slow
GHA debugging and allow us to write builds and integration tests exactly
as they run in CI. However, there's a tough problem we have to work
around to achieve both local builds and good caching for fast CI builds.

Most of us work on Mac laptops, so if we do local builds, we have to do
them in some kind of Linux environment. The easiest way to do this
without a lot of custom tooling is by containerizing builds with Docker.
This is what I've been working toward in past PRs, and it works right
now.

However, I've come to realize that in CI, the caching properties of
these images I made are very different. The [`RUN --mount=type=cache`
args I
added](https://github.com/onefootprint/monorepo/blob/826d794b82d28259d0852cdc9949384ed7588fa7/backend/images/api/Dockerfile#L10-L11)
maintain a cache local to the Docker builder that is not exported with
the regular Docker layer cache, which is controlled by the
[`--cache-from/--cache-to`
flags](https://github.com/onefootprint/monorepo/blob/826d794b82d28259d0852cdc9949384ed7588fa7/.github/workflows/backend_ci.yml#L80-L81).
This means that these Docker builds are quick locally, but slow in CI
because dependency crates are built from scratch every time.

Getting Cargo to play nicely with Docker's layer caching is a hard
problem. There's a very good document
[here](https://hackmd.io/@kobzol/S17NS71bh) that explains some of the
solutions to this problem.

I played around with a few of the current solutions, and the one that
seems most promising with minimal added complexity or finicky config is
[cargo-chef](https://github.com/LukeMathWalker/cargo-chef).

cargo chef runs in two parts. First, it generates a `recipe.json` plan
for the dependencies to build. Then, it builds all the dependencies. As
long as the content digest of `recipe.json` remains the same
(effectively `Cargo.lock`) and the Dockerfile remains the same, the
build of the dependencies will be cached as a Docker layer. We store
this cache in the GHA cache, and mostly for now rely on the fact that
the 10gb limit is a soft limit that is only enforced several times
throughout the day (seemingly).

### Notes on `Swatinem/rust-cache`
https://github.com/Swatinem/rust-cache is the predominant way to cache
in native GHA, and we use it too. It tries to clean up the `target`
directory of non-dependency artifacts. Unfortunately this is a
GHA-specific Typescript that we can't run in Docker. Getting this cache
to work inside Docker require bind mounting the cache directory at
build-time and preparing it for saving properly, similar to [how it has
been historically done in our image
builds](https://github.com/onefootprint/monorepo/blob/fc65d9f17f1bd7a1e79473fccf120cadb99359da/.github/workflows/backend_build.yml#L60-L117).
For cargo inside Docker to actually use the cache, the Rust environment
needs to be identical both inside the container and in GHA. We'd be
duplicating build config and we would have a finicky binding between the
two environments. In my opinion, using `cargo-chef` until there's a
better solution in https://github.com/rust-lang/cargo/issues/2644 is
simpler and easier to maintain.


## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->