[BE-269] Replace null scoped vault status with explicit `ObStatus::None` (#9976)
## Describe your changes
We'll eventually start to have users that have been through a workflow
that have an explicit `None` status. For ex, if a user goes through a
`skip_kyc` playbook, we will use status None.

In short, `None` means "rules not executed" while` Pass` means "rules
executed and no rules matched, so we defaulted to pass."

## Caveats
This makes a few breaking changes to APIs by adding a new enum variant.
I've audited all of them. Most of them are dashboard-facing, and luckily
the dashboard has (mostly) already come up with its own in-memory
representation of "none" - it mapped this null status to an explicit
`'none'` string, which works perfectly with this migration. See old
logic
[here](https://github.com/onefootprint/monorepo/blob/9ae8bcbaec0354cde5630fa098825f5c1d96bf15/frontend/packages/types/src/data/entity.ts#L149-L154).

The only somewhat concerning API change is in the `GET /users/{fp_id}`
API. We will now start returning `"status": "none"` instead of
`"status": null`.

Luckily, this is a preview API! ~so its use is very limited.~ I cannot
really imagine a case in which a tenant is looking explicitly for `null`
status on users since this was only really possible for vaults created
via API (which generally wouldn't be necessary).

Edit: unfortunately, lots of tenants do have access to this preview API.
Only a few use consistently, with a others seeming to have occasional
uses of the API
([metrics](https://ui.honeycomb.io/footprint-2e/environments/prod/datasets/fpc-api/result/zg5Yt3Dbbr9?hideCompare)).
For this reason, I've decided to be safe I'll just serialize
`Some(OnboardingStatus::None)` as `null` for now. I'll advertise that
we're making this change to tenants and planning to switch this new
serialization

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->