Fix workflow status tracking for skip_kyb KYB playbooks (#10014)
## Describe your changes
Trayd is currently the only customer that has a playbook with
skip_kyb=true. I think we made a few bandaid patches to support them a
while back, so this cleans up a few of those :)

First, when skip_kyb was true, we would skip straight from
`KybAwaitingBoKyc` to complete, without going through decisioning. this
differs from our KYC workflows, which do pass through _both_ the vendor
calls and decisioning states even when skip_kyc is true. Now, we at
least just jump to the kyb decisioning state

When in the KYB decisioning state, we'll evaluate rules. If it's a skip
KYB playbook, you'll have no rules so the rules engine will output that
no rules executed. With some recent changes in this stack, that
automagically sets the business's workflow's status to none and the
business's status to none! No more need for this special case handling
with the very custom `ScopedVault::clear_business_status` function

## Caveats
This is technically a breaking change to what trayd is seeing for their
skip KYB playbook. But i cannot imagine they are reading the current
nonsensical workflow status returned in the validate API. They don't
have webhooks set up either so no problem to start returning a none
status

## Test Plan