Use safe shell settings in the enclave cloud-init and fix bug (#10015)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
Currently, as noted in the TODO, we're running the startup script as a
fire & forget. This adds safe shell settings to improve reliability &
debugability. Since we didn't have these flags set, we had a latent bug
in the script -- an undefined variable `$USER`. Since `usermod` command
failed without consequence, I removed it. For the `chown` calls, I
switched these to set the owner to root, since that's the user that
ultimately runs the files.

Another issue I found at the same time is that the tailscale connect
script runs in a hot loop, rather than once at startup, so systemd marks
it as failed. In a future PR I'll make this a oneshot systemd service
with retries.

It would be great to get these kinds of logs in Datadog -- we should
install the agent on the enclave instances soon.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->