[FP-7755] Add GET /businesses/{} API (#10035)
## Describe your changes
Finally adding this counterpart to the `GET /users/{}` API :)

## Caveats
There are still some user-facing APIs (client token, upload, search,
labels + tags, etc) who don't have a business analog. I'm okay building
these as tenants request them - it's nice to not have to support them if
they're not in use.

Last week, we added a status `none` as a possible option in the `GET
/users/{}` response, which required some extensive auditing of which
tenants are using the API and how they use it. It was nice to not have
to do the same for a `GET /businesses/{}` response too

## Test Plan