Use canonical images in local integration tests (#9965)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
Now, we start to realize the benefit of building images pre-merge in CI.
We remove the duplicate build definition for integration tests and use
the pre-built images.

In this PR I threw in a build of `footprint-dr` to the integration test
image. This will enable us to write new integration tests that use that
CLI.

To develop new non-pytest integration tests integration tests locally,
you'd do something like this:
```
docker compose up -d
docker compose run integration_tests ./ci/scripts/my_new_test.sh
```

I intend to do this with footprint-dr, running with our third-party AWS
test account credentials.

Remote integration tests will come in future PR.

On the topic of caching/build time:
Cached API release image builds take about 5 minutes, but integration
tests take only 1 minute 30 seconds to 2 minutes. But now we're actually
running integration tests on the same build definition with the same
system dependencies, etc. So hopefully the benefit is worth the slight
difference in time. Part of the increase in build time is because we're
now only building the API server in release mode, whereas before we were
building it in debug mode on the fly. Pros and cons...

I'm not completely sold on cargo-chef yet, since I see that some API
dependencies are being rebuilt still. It works great on the enclave
images though. More to experiment with in the future but I want to move
on for now, and the build times are okay.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->