[BE-19] Always put user into manual review after document workflow (#10000)
## Describe your changes
We've always felt uneasy about the behavior in the document workflows
that implicitly ran rules from the last playbook the user onboarded
onto.

I've felt this for a while, but after the composer shadowing session, it
made it very clear that customers expect when requesting an ad-hoc
document that a user is put back into the manual review queue. This
finally changes this behavior - instead of running rules at all on these
ad-hoc document workflows, we just put the user back into manual review.

## Caveats
We still make a timeline event for the `OnboardingDecision` when the
user finishes the ad-hoc document workflow, and this timeline event kind
of erroneously mentions a playbook. I think would be better to omit the
playbook entirely...

## Test Plan


https://github.com/onefootprint/monorepo/assets/21066341/a2f170e1-4ab7-40ea-b8b3-03dc716127d9