Simplify Rust build caching (#10046)
## Describe your changes
cargo-chef is a pretty complicated dependency IMO and it also adds a lot
of complexity to our image definitions. Even though it's recommended in
the docs for clux/muslrust, I think what I'm proposing in this PR is
simpler and seems to work a little better. I found that this proposed
method more reliably caches all API dependencies and it runs slightly
faster.

In this PR I:
- Replace cargo-chef with native Docker `cache` build mounts that are
copied in/out using of Docker
`reproducible-containers/buildkit-cache-dance@v3.1.0` and cleaned up
using our trusty `Swatinem/rust-cache`.
- Build the API server in debug mode pre-merge to shave off a couple
minutes from the build.
- Drive-by improvements:
- Remove debug symbols from debug builds. This may improve debug build
times, but principally I'm doing it here to keep debug cache sizes down.
- Disable incremental builds in Docker, which `Swatinem/rust-cache` does
to keep cache size down.

The result is an API image that builds in <5 minutes with a warm cache
on PRs.

## Caveats


## Test Plan