[fe] Start reading vault data errors from error context (#10080)
## Describe your changes
The backend is going to start mandating that the error `message` is
always a string, so the per-field validation errors are now going to be
in the `context` field. All of the vaulting APIs (`PATCH
/hosted/user/vault`, `PATCH /hosted/business/vault`, `PATCH
/entities/<>/vault`) used to return an error that had a JSON object in
the error's `message`. This migrates the clients of all of those APIs to
read the error's `context` instead of the `message`.

And, the backend is removing this `error` envelope around the error
response (see the PRs below this for context).

## Caveats


## Test Plan
Locally, ran a future version of the backend that has the modern error
payload with a string message. Tested that error messages from the
backend propagate to the frontend, both in idv and in the components SDK
demo:


![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/0e4f5863-900f-4f7d-9900-70c8b943a682.png)


![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/2b31539b-abeb-4214-b40c-bf8aa63f990e.png)